repos/services/model-health/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.32.3
PyYAML==6.0.2
